WT-6446 rename *.i into *_inline.h files
WiredTiger use of XXX.i as the file name for inlined functions is a little odd,
generally the XXX.i suffix is reserved for files that do not require pre-processing,
and in the case of MSVC, it's a Visual Studio Intermediate file that contains
the output of the preprocessor.
