+++ src/include/bitstring_inline.h
