+++ src/include/block_inline.h
