+++ src/include/buf_inline.h
