+++ src/include/cursor_inline.h
