+++ src/include/intpack_inline.h
