+++ src/include/log_inline.h
