+++ src/include/misc_inline.h
