+++ src/include/os_fstream_inline.h
