+++ src/include/serial_inline.h
