+++ src/include/timestamp_inline.h
