+++ src/include/wt_internal.h
@@ -411,31 +411,31 @@ typedef uint64_t wt_timestamp_t;
 #endif
 #include "verify_build.h"
 
-#include "cache.i"   /* required by misc.i */
-#include "ctype.i"   /* required by packing.i */
-#include "intpack.i" /* required by cell.i, packing.i */
-#include "misc.i"    /* required by mutex.i */
+#include "cache_inline.h"   /* required by misc_inline.h */
+#include "ctype_inline.h"   /* required by packing_inline.h */
+#include "intpack_inline.h" /* required by cell_inline.h, packing_inline.h */
+#include "misc_inline.h"    /* required by mutex_inline.h */
 
-#include "buf.i"       /* required by cell.i */
-#include "timestamp.i" /* required by btree.i */
-#include "cell.i"      /* required by btree.i */
-#include "mutex.i"     /* required by btree.i */
-#include "txn.i"       /* required by btree.i */
+#include "buf_inline.h"       /* required by cell_inline.h */
+#include "timestamp_inline.h" /* required by btree_inline.h */
+#include "cell_inline.h"      /* required by btree_inline.h */
+#include "mutex_inline.h"     /* required by btree_inline.h */
+#include "txn_inline.h"       /* required by btree_inline.h */
 
-#include "bitstring.i"
-#include "block.i"
-#include "btree.i" /* required by cursor.i */
-#include "btree_cmp.i"
-#include "column.i"
-#include "cursor.i"
-#include "log.i"
-#include "os_fhandle.i"
-#include "os_fs.i"
-#include "os_fstream.i"
-#include "packing.i"
-#include "reconcile.i"
-#include "serial.i"
-#include "time.i"
+#include "bitstring_inline.h"
+#include "block_inline.h"
+#include "btree_inline.h" /* required by cursor_inline.h */
+#include "btree_cmp_inline.h"
+#include "column_inline.h"
+#include "cursor_inline.h"
+#include "log_inline.h"
+#include "os_fhandle_inline.h"
+#include "os_fs_inline.h"
+#include "os_fstream_inline.h"
+#include "packing_inline.h"
+#include "reconcile_inline.h"
+#include "serial_inline.h"
+#include "time_inline.h"
 
 #if defined(__cplusplus)
 }
